CHG: Use BigDecimal for monetary amounts instead of double

+++ banco/src/main/java/app/application/adapters/persistence/sql/entities/BankAccountEntity.java
@@ -14,6 +14,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -48,7 +49,7 @@ public class BankAccountEntity {
   private Currency currency;
 
   @Column(name = "current_balance")
-  private Double currentBalance;
+  private BigDecimal currentBalance;
 
   @Column(name = "opening_date")
   private LocalDate openingDate;
